[refactor] Move packages to another place
