--- app/src/main/java/com/makentoshe/habrachan/di/MainFlowFragmentScope.kt
@@ -1,5 +1,0 @@
-package com.makentoshe.habrachan.di
-
-@Target(AnnotationTarget.CLASS)
-@Retention(AnnotationRetention.RUNTIME)
-annotation class MainFlowFragmentScope
