+++ app/src/main/java/com/makentoshe/habrachan/di/main/MainFlowFragmentScope.kt
@@ -1,0 +1,5 @@
+package com.makentoshe.habrachan.di.main
+
+@Target(AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.RUNTIME)
+annotation class MainFlowFragmentScope
