+++ app/src/main/java/com/makentoshe/habrachan/di/main/posts/PostsFragmentScope.kt
@@ -1,0 +1,5 @@
+package com.makentoshe.habrachan.di.main.posts
+
+@Target(AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.RUNTIME)
+annotation class PostsFragmentScope
