--- app/src/main/java/com/makentoshe/habrachan/di/posts/PostsFragmentScope.kt
@@ -1,5 +1,0 @@
-package com.makentoshe.habrachan.di.posts
-
-@Target(AnnotationTarget.CLASS)
-@Retention(AnnotationRetention.RUNTIME)
-annotation class PostsFragmentScope
