--- app/src/main/java/com/makentoshe/habrachan/model/account/AccountFragmentScreen.kt
@@ -1,10 +1,0 @@
-package com.makentoshe.habrachan.model.account
-
-import androidx.fragment.app.Fragment
-import com.makentoshe.habrachan.common.navigation.Screen
-import com.makentoshe.habrachan.view.account.AccountFragment
-
-class AccountScreen : Screen() {
-    override val fragment: Fragment
-        get() = AccountFragment()
-}
