+++ app/src/main/java/com/makentoshe/habrachan/model/main/MainFlowScreen.kt
@@ -1,0 +1,10 @@
+package com.makentoshe.habrachan.model.main
+
+import androidx.fragment.app.Fragment
+import com.makentoshe.habrachan.common.navigation.Screen
+import com.makentoshe.habrachan.view.main.MainFlowFragment
+
+class MainFlowScreen : Screen() {
+    override val fragment: Fragment
+        get() = MainFlowFragment()
+}
