+++ app/src/main/java/com/makentoshe/habrachan/model/main/account/AccountFragmentScreen.kt
@@ -1,0 +1,10 @@
+package com.makentoshe.habrachan.model.main.account
+
+import androidx.fragment.app.Fragment
+import com.makentoshe.habrachan.common.navigation.Screen
+import com.makentoshe.habrachan.view.main.account.AccountFragment
+
+class AccountScreen : Screen() {
+    override val fragment: Fragment
+        get() = AccountFragment()
+}
