--- app/src/main/java/com/makentoshe/habrachan/model/posts/PostsResponseCache.kt
@@ -1,19 +1,0 @@
-package com.makentoshe.habrachan.model.posts
-
-import com.makentoshe.habrachan.common.cache.Cache
-import com.makentoshe.habrachan.common.cache.CacheStorage
-import com.makentoshe.habrachan.common.network.request.GetPostsRequest
-import com.makentoshe.habrachan.common.entity.posts.PostsResponse
-
-class PostsResponseCache(
-    private val storage: CacheStorage<GetPostsRequest, PostsResponse>
-) : Cache<GetPostsRequest, PostsResponse> {
-
-    override fun get(k: GetPostsRequest): PostsResponse? {
-        return storage.get(k)
-    }
-
-    override fun set(k: GetPostsRequest, v: PostsResponse) {
-        storage.set(k, v)
-    }
-}
