+++ app/src/main/java/com/makentoshe/habrachan/ui/main/posts/PostsFragmentUi.kt
@@ -1,0 +1,12 @@
+package com.makentoshe.habrachan.ui.main.posts
+
+import android.content.Context
+import android.view.LayoutInflater
+import android.view.View
+import com.makentoshe.habrachan.R
+
+class PostsFragmentUi {
+    fun createView(context: Context): View {
+        return LayoutInflater.from(context).inflate(R.layout.main_posts_fragment, null, false)
+    }
+}
