--- app/src/main/java/com/makentoshe/habrachan/ui/posts/PostsFragmentUi.kt
@@ -1,12 +1,0 @@
-package com.makentoshe.habrachan.ui.posts
-
-import android.content.Context
-import android.view.LayoutInflater
-import android.view.View
-import com.makentoshe.habrachan.R
-
-class PostsFragmentUi {
-    fun createView(context: Context): View {
-        return LayoutInflater.from(context).inflate(R.layout.main_posts_fragment, null, false)
-    }
-}
