--- app/src/main/java/com/makentoshe/habrachan/view/menu/MenuFragment.kt
@@ -1,16 +1,0 @@
-package com.makentoshe.habrachan.view.menu
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.TextView
-import androidx.fragment.app.Fragment
-
-class MenuFragment : Fragment() {
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        return TextView(requireActivity()).apply {
-            text = "Menu fragment"
-        }
-    }
-}
